Revert most of the changes from 16b76a4
At the time, Parsoid only emitted the figcaption if a caption was
present.  So, having the border set on the figcaption was leading to
some incomplete borders on figures.

However, since I856ee962b70cef1f8d49652396ea5264e11a8ade,
that's no longer the case.  We can now switch back to displaying the
figcaption as a table-caption and don't have to rely on the hacky
`width: 1px` to have the figcaption width match the media in the figure
(see https://stackoverflow.com/a/6536025).

One problem with the 1px hack is that in T241392, TimedMediaHandler sets
the max-width of a video to 100%, which shrinks it down to that 1px.

The change in display for the figcaption necessitates a bit of trickery
to get the magnify clip in the right place.  At the time of the reverted
commit, css for the magnify clip wasn't present.

Bug: T272186
Bug: T266149
Change-Id: If283c1ab200ca1049343cdac759d7a2b78306227