objectcache: Fix non-numeric string check in HashBagOStuff for PHP 8
PHP 8 changes the handling for string and int comparison.
These non-numeric strings are no longer silently converted to int 0,
which previously allowed for the comparison to become false without
actually checking against non-integer values.
https://wiki.php.net/rfc/saner-numeric-strings

This fixes a unit test failure on PHP 8:
> HashBagOStuffTest::testConstructBadType
> Failed asserting that "InvalidArgumentException" is thrown.

This fails because the following new warning is encountered
instead, which PHPUnit turned into an exception:

> PHP Warning: A non-numeric value encountered

Bug: T248925
Change-Id: I6f75080ac8d549b815932527accb2fd5129d5775