Consistently order change tags
When tags are shown on history pages, recent changes, and so on,
they will now be shown in the same order as the result of
listDefinedTags(), rather than by their ctd_id (corresponding
to the order in which they were first used on a given wiki).

This allows extensions using the ListDefinedTags hook to have their
tags displayed in a specific order consistently on every wiki.

Bug: T272900
Change-Id: I4c85cb840c1fac52f5ab268beee9e6a7fd20b91e