DatabaseUpdater: Remove renameIndexes() in favor of renameIndex()
This was introduced in current master 1.36, after discussion in
I17028bb. It's never been part of a stable release and so is skipping
deprecation process.

The issue discussed in I17028bb is still there and valid (I believe),
but this method alone cannot solve that problem and so there's no point
in keeping it per discussion in I6751f0f

Change-Id: I673b9a45b29ed88cf677412978dfa8b7c754405e