Call TestCase::fail() more carefully in MockHttpTrait
If we directly pass [ TestCase::class, 'fail' ] as the callback for a
mock method, then the method will be called with whatever arguments the
caller tried to pass to the original method. Those arguments may not be
compatible with TestCase::fail()’s signature, which requires the first
argument to be a string (or absent). Use a custom callback first which
throws away any arguments before calling TestCase::fail().

Bug: T272998
Change-Id: I3514d3d29355a55a5ca10ccc22981e3972347f35