mediawiki.base: Fix incorrect 'this' value in mw.msg()
Follows-up r75490 (f8eca3d823), which introduced the `mw.msg` method
but passing `mw.message` instead of `mw` as the "this" value when
calling the mw.message() method.

This is harmless, and being a static method it is unlikely to
(and indeed doesn't) make use of it, but was incorrect
nonetheless.

Change-Id: I26348210965c7f75f88be561376a32751d41bd39