objectcache: Always enable debug mode in BagOStuff
BagOStuff has a debug mode because MemcachedClient has a debug mode,
because Ryan T. Dean put one in there in 2003. It's not strictly
necessary now that we have log levels.

This means that a few existing $this->logger->debug() calls are now on
the same effective level as $this->debug() calls, they can't be enabled
separately. But that seems inconsequential.

Change-Id: I4d1cb9153559b29468b6ca530bbc56e2b35d1721