[docker] Set user/group mapping for all host OS, not just Linux
Apparently Docker for Mac (I assume also Docker for Windows) has changed
and we now need to set the host user/group mapping to avoid permissions
problems in the running container. As a result, we don't need to make
Linux users create an override file anymore.

Also, update the docker-compose.yml to reference the `.env` file,
apparently this explicit reference is required for those values to be read.

Bug: T273218
Change-Id: Iaadce5d27d6e973ad02afd5b069646e95b5de0aa