objectcache: make MediumSpecificBagOStuff::updateOpStats more robust
Make the method accept either an integer-indexed list of keys or a map
of keys to the send/receive payload size information tuple.

Make sure that no BagOStuff subclasses are passing in $keys directly
from the call to the public BagOStuff methods; these allow keys in the
form of a list or a map of arbitrary strings to keys.

Change-Id: I9687d25a4dd1c7b4b304f9fd543cc0a26a595962