API: Micro-optimize ApiResult::isMetadataKey()
This is an extremely hot function, appearing at the top or second in a
few XHGUI profiles I ran when sorted by self wall time.

We can micro-optimize by switching the substr() function call to check
the 0th character of the string.

In my benchmarking of 1,000,000 loops (see P14054), this went from a
total time of 784.20ms to 651.67ms.

A further optimization could remove checking whether the string is empty
but that would be a behavior change and while I'm pretty sure all API
keys are non-zero length, I'm not 100% sure.

Change-Id: I368cd3b526edeb247a6ebda7420897a51357407d