Fix mismatching type hints in ChangesListSpecialPageTest
Some of these parameters are marked as optional and nullable,
but can never be null (the test would fail then), and are
never missing.

While I'm here I also add more strict type hints for
parameters and return types. Note this is only done in tests
and impossible to cause trouble in production.

Change-Id: I1c66325c97d5ea86d9822e789a36542faad1272b