ApiLoginTest: Remove testGotCookie that's been broken for 10 years
For a decade (since https://phabricator.wikimedia.org/rMWa11cf96d0688)
this test has not been executed. It's safe to assume that it's useless
now and this code is better deleted.

It still fails if re-enabled and it seems to me the reason for its
brokenness is still there, so it's simply not fixable.

Bug: T273603
Change-Id: I6887cb88ae0a2d6665b018246215855498a7ff42