Re-enable the lone test in ApiPurgeTest
This was disabled 10 years ago because Jenkins was not happy.
We should reenable it and see whether it's happy now, otherwise
the test should be deleted.

Use MediaWikiTestCase::getExistingTestPage() to make sure the page
exists which is another reason given for skipping the test even if
not broken in whole.

Change the test name to more correct 'testPurgePage'

See https://phabricator.wikimedia.org/rMW7a0353b01080

Bug: T273603
Change-Id: Ief2017751256c683c8e4e4df375d5a82bcb4849c