Fix documentation of IDatabase::selectField()
I found code in an extension that actually does this and
provides an array of field names (obviously with a single
field name only). This might be an "unintentional feature".
Still I believe it's a good idea to support it, because all
other methods support it as well. This makes selectField less
of an outlier.

Change-Id: Ie63b06896b59dcaa629f720f98f28943674dc8d7