Micro-optimize ApiResult::isMetadataKey even more
Since this is a very hot method, it must be really quick. ord() is
useful here, because it discards everything after the first character
and it works with integers.

A quick benchmark on 3v4l reveals that it's consistently faster than
the previoust code on all PHP versions, taking something between
38% and 71% of the current runtime, with an average of 50%.

Follow-up: I368cd3b526edeb247a6ebda7420897a51357407d
Change-Id: I0b471dafd645aeb11ecdc68915c28aff93f809d9