Change color of "You have a new message" to align with the WikimediaUI
This is the really old way of notifying users they have a message in
their talk page and still being shown to logged-out users. The color and
border is pretty old and doesn't follow the UI guidelines.

This change makes it similar to logged-in users (color wise) by removing
the border, adding a border radius and changing the background color
to #fc3 (the Wikimedia color). The change is basically invisible to the
naked eye.

The whole thing needs a rework which is being discussed in T266511

Bug: T278105
Change-Id: I30eea1278b3116debcd06a76ae16560898259b13