rebuildLocalisationCache.php: handle killed children
Before:

If a child terminated due to a signal (for example, being killed by
the kernel OOM killer), that would go unnoticed in the parent process
and the parent would ultimately exit with a status of 0.

After:

pcntl_wifexited() and pcntl_wifsignaled() are now used to determine if
a child process terminated normally or via signal.  In the case of
normal termination with a non-zero exit status or termination by
signal, a message is printed.  In either case the parent will exit
non-zero.

Change-Id: I646ac6cbd8cd653553718d55055cad7cbd4b6bc1