MemoizedCallable: Disallow memoizing of unnamed closures
Using a random id as the cache key is useful in that it ensures
we don't poison ourselves with results from unrelated closures,
but it also prevents any use of APCu caching beyond the current
web request which is kind of what it is for.

There is a subclass inside the test case that stores results
in-process with an array instead, but for things like that we
generally use in-object caching of class instances, or
HashBagOStuff::getWithSet, or MapCacheLRU::getWithSet.

Change-Id: I264ad87af25dda4cde0c71aad83b30c314f8cd8d