API: Optimize hot code paths in ApiResult by re-arranging
This does not change what this code does in any way. All it
does is re-arranging the code paths from most likely to least
likely. For example:

* It's much more likely that an element in an API result is a
string, and much less likely it's an array. This change
increases the chance that the long if-elseif ends after the
first check, and rarely executes the other checks.

* There are not many arrays in an API result, but many, many
more scalar values. Doing the is_array() check first avoids
many, many calls of the hot isMetadataKey() function.

This makes quite a difference in my benchmarks.

Change-Id: I47c1525bdedbc8ca26e10ea7480715a7ecd7d5b2