registration: Allow reusing cached metadata between wikis
ExtensionRegistry caches components of load queues in APCu, varying the cache
keys on its own internal version, the MediaWiki version, the set of extensions
to be loaded etc. - in short, everything that may affect their contents.
However, it then uses makeKey() to build the keys, which scopes the keys to the
local wiki, making it impossible to reuse cached queues between wikis when the
parameters match.

On installs with a large number of wikis, this causes ExtensionRegistry to fill
APCu almost immediately after startup, resulting in a constant expunge cycle. At
Fandom, we implemented a patch[1] to allow reusing cached load queues between
different wikis by scoping the cache keys the global keyspace, instead of
generating wiki-local keys. This dramatically reduced the amount of ER cache
misses we observed, going from ~10% of total request time as reported by ArcLamp
to <0.5%.

[1] https://github.com/Wikia/mediawiki/pull/14

Bug: T274648
Change-Id: I3bee5d01399dcc965ec33a0d573ead144f1144a7