mediawiki.action.edit: Set desktop/mobile target
This module is queued by EditPage for the action=edit interface.

This generic module contains almost nothing and, while this is
disabled on mobile today, its dependency (oojs-ui-core) and
sibling modules (edit.editWarning, and edit.watchlistExpiry)
are already explicitly enabled for mobile on EditPage today, so
most of what one would expect from this module is already
working as expected.

What will start working for mobile action=edit:

1. Edit summmary character limit.
   If you have typed more than 900 of the 1000 allowed characters,
   a countdown from 100 will display in grey text on the right
   side within the input field.

2. Edit stashing ajax.
   Some time before the user submits the form, an ajax background
   request will start pre-parsing the wikitext to reduce latency
   and improve edit timing.

Having said all that, none of this matters in all likelihood since
we don't actually expose action=edit on mobile to users with
Grade A browsers (where this JS would run). On mobile, the
action=edit interface is generally only seen when using a Grade C
browser or when JS failed to load or was disabled.

Enabling this module helps unblock T127268 (to no longer suggest
that this code is unsafe for mobile) and T274688 (to fold the
the other edit* modules that are already enabled on mobile into
this one).

Bug: T274688
Bug: T127268
Change-Id: I953ea0a60ab7cacd174e09fa52288f3e1d811611