PostgresUpdater: Move `bot_passwords` table addition down
This should have no effect, but it helps in making this file
consistent with  Mysql/Sqlite updaters.

Table was introduced in 1.27 (Iaa4015e) and was added in
'new tables' section, because then Postgres separated them
per patch type instead of per release.

Separating them per release (as MySql/Sqlit do) is more logical
and it's even more important now that we are not keeping them
indefinitely.

Bug: T272199
Change-Id: I1fa6d81c9903fb6188f00b791794b3199a0a5aa3