Force unique ids for cloner-generated elements
Cloner elements are created using the data-template attribute in
the parent element. The first template use will result in unique
ids but subsequent uses reuse the template ids.

Use the cloner count to make each id unique whenever the cloner
generates new elements. It is safe to do so because ooui-php
is not used for any other sort of data management, only to distinguish
between unique elements. By appending onto the id, there will also
be no overlap with ooui elements generated elsewhere on the form.

Bug: T274533
Change-Id: I96bb093a83303de7d65772f9b807ed0c735005b2