Implement UserIdentityValue::__toString()
Currently in some places we pass User as a message parameter.
This works cause User::__toString() returns the user name,
which in general makes sense. Now that we are replacing User
with UserIdentityValue, the UIV instances will start flowing
around in the system more, and some UI messages will definitely
break.

This is not an ideal solution, we probably should have special
ParamType::USER in the Message infrastructure, and have a special
formatter for this type, but it would be extraordinary hard to track
down all the place where this might be needed, so perhaps let's
emulate User behaviour in UserIdentityValue.

Change-Id: I5c6a25234d1599178d08212fb2d2f197b433b164