Don't accept empty option group names for dropdown elements
When this code parses a string that looks like this:

*
** A
** B

It creates an option group with 2 options, but the name
of the group is an empty string. This makes the OOUI
DropdownInputWidget fail later.

Since the empty group name is useless anyway, drop it. This
makes the code behave as if the parsed string looked like
this:

* A
* B

Live example for the issue:
https://el.wikipedia.org/wiki/MediaWiki:Protect-dropdown

Bug: T275125
Change-Id: I780c1be27740b0ed3b35aa569b5a528112d7238f