Make RememberMe authentication behavior configurable.
RememberMeAuthentication request shows a "remember me" checkbox that,
when checked, will extend the amount of time before the authenticated
session expires. It was previously not possible to extend the session
when the login form is skipped, which happens when there are no
non-skippable fields on the form. This patch introduces a configuration
variable, $wgRememberMe. Valid values are:

- CHOOSE_REMEMBER: the user may be able to choose whether to be
remembered or not (depends upon whether login form is skipped)

- FORCE_CHOOSE_REMEMBER: the user will be able to choose whether to be
remembered or not (forces login form not to be skipped)

- ALWAYS_REMEMBER: the authenticated session will always be extended

- NEVER_REMEMBER: the authenticated session will never be extended

Default behavior is identical to what it was before adding this
functionality.

Bug: T265263
Change-Id: I779aae3c1b96b380b50092245f616219088b038d