Remove unused patch-linktables.sql
This file was never wired to update.php since its introduction in 2004
and was added to documentation as something you can run if you want to.

That documentation was removed in 2005 and since then, there is no
mention of this file anywhere and it's clearly unfit to run in current
schema.

The commit that introduces it: https://phabricator.wikimedia.org/rMW59c6e92429c
The commit that removes documentation: https://phabricator.wikimedia.org/rMW10b01d61eae280

Change-Id: Ib98f2275f9dfba0870d1638274ddf05ecf50b5b1