Add Vue.createMwApp(), to help with Vue 2->3 migration
This will allow code that mounts components to be written in a way that
will keep working across the Vue 2->3 migration:

    var RootComponent = require( './RootComponent.vue' ),
        store = require( './store.js' ); // if using Vuex
    Vue.createMwApp( RootComponent )
        .use( store ) // if using Vuex
        .mount( '#foo' );

While we're still using Vue 2, createMwApp() will return something that
pretends to be an app object, with support for .use() and .mount().
Once we move to Vue 3, createMwApp() will remain, as a wrapper around
Vue.createApp() that adds in the i18n plugin and the error logger.

Migrating to this API won't be required immediately, since the change
upgrading Vue from v2 to v3-compat will include a compatibility wrapper
for the old calling style.

Bug: T251974
Change-Id: Ib3ef9b88547b43c566abb007cde08a27baf2b7a4