Remove unused patch-oldestindex.sql
This file was never wired to update.php since its introduction in 2003
(r1348) and was added to documentation as something you can run if you
want to.

That documentation was removed in 2005 (r23242) and since then, there is no
mention of this file anywhere and it's clearly unfit to run in current
schema.

Change-Id: If2aa9900e33f7a728ac47bf9cc7610ba62409052