Don't worry about something before when armoring french spaces
We lost some insight in c44a395 because we're no longer analysing the
entire dom as a serialized string, but instead running our regexp on
individual text nodes.

This patch as written here just allows for the space to be at the start
of the text node.  However, some git spelunking shows that in 9dc65ef,
the condition for there being a non-whitespace character previous to the
space was only because armoring French spacing happened before
doBlockLevels and wanted to protect indent pre's.

That's certainly not the case anymore, so we can probably get away with
dropping the condition altogether now.

Bug: T275918
Change-Id: I654a09b0f98937379b9fad3f325134ead7f2d8a6