Control whether a log is created when accounts are autocreated
When autocreation is forced by another user, that action is logged
already and attirbuted to the correct actor (e.g. the sysop who
forcibly autocreated the account). The 'autocreate' log should not
be created in this scenario because it would incorrectly attribute
an action to the subject of the autocreation, when they indeed had
not taken any action themself. Collateral damage is explaiend in
T275979.

This patch also corrects the wording of the method's documentation

Bug: T269494
Change-Id: I9c771a6a8bed7711508f43b793c295cd30e4ea6b