Re-introduce UserIdentity::getUserId - removed in a violation of SIP
UserIdentity::getUserId was removed by 493105109e
in a violation of [[:mw:SIP]], as the removing developers
did not make sure this change does not affect Wikimedia wikis,
and as getUserId was never hard-deprecated, in fact,
it was _recommended to be used_ by UserIdentity::getId's
soft deprecation.

Reinstante the removed code to unbreak ie. SecurePoll.

Bug: T275482
Change-Id: I306c33165b2bf08602d2bbc60782e0a493405797