Simplify PageArchive::__construct()
Use typehinting instead of inefficient manual handling.

It's inefficient because if you pass anything that's not null,
(and not Title), there'd be no error during the construction but
an error will definitely manifest whenever a title method is
called on whatever thing happens to be held in the $title prop.

Change-Id: I5e34a3ed28ad5f892ee6280b3acf174112e6df9e