phpunit: Fix implementation of anythingBut
As I wrote in Ie9d5d7e53faee09b7fe63f9387f72dc87411c24d,
Assert::matches:
- Only accepts strings (and not arbitrary values)
- And it does so because the comparison uses format strings (à la
  printf), which is not what we want here.

This currently works for method names like the docs suggest, but it
fails with basically anything that is not a string without '%' chars.

Instead, use identicalTo() which performs === comparisons. Note that
using equalTo() (which uses `==`) would perhaps be more consistent
with the PHPUnit API, but then we'd have the usual problems with weak
comparisons. A separate implementation (or method parameter) which uses
== can be added later if someone really finds it necessary.

Change-Id: I8444129198182f5eccc0a7282abd609b2f763299