Add Wayuu to Names.php
The autonym is different from the current definition in translatewiki.
The name "Wayúu" was added to translatewiki configuration in 2009,
and it's unclear what is its source. All the sources that I can
find say that autonym is "wayuunaiki", for example Ethnologue,
Diccionario básico ilustrado wayuunaiki-español español-wayuunaiki
(Editorial Fundación para el Desarrollo de los Pueblos Marginados,
Bogotá, D.C., 2005), and the Incubator Wikipedia in this language.
I also verified this with User:Maor_X, who is familiar
with the Wayuu culture.

Spanish is added as fallback, as it was in translatewiki, and it
generally makes sense, because the language is spoken in
Spanish-speaking countries and many speakers are literate
in Spanish.

Bug: T276745
Change-Id: I04d698e7c27d96f7e58c62f6f71e76142a30630d