Add Maintenance::waitForReplication()
Common utility method for maintenance scripts that's a little more
clever than LBFactory::waitForReplication(). Previously it was
included in Maintenance::commitTransaction but that logs an error
when there is no uncommitted change, and one might want to commit
more often and wait for replicas to catch up only after some amount
of commits.

Change-Id: I3394536eea01eb982a4a2033fd2062bc67f6bdc1