Message: update docs to discourage wfMessage usage
Update the phpdoc of the Message class to encourage object creation
via MessageLocalizer::msg() instead of wfMessage() as the latter
results in calling to the global request context during message
parsing, which leads to unintuitive errors when sessions are not
available (in ResourceLoader callbacks or early hooks like
LocalUserCreated).

Also drop the documentation about the (long since removed) wfMsg*
methods.

Change-Id: Ie8c5c5e055cb8ff79d47d4c129df56f719e1cba6