Escape mediastatistics-header-* messages on Special:NewFiles
The mediastatistics-header-* messages used on Special:NewFiles currently
allow raw HTML as raw options labels are output as raw HTML in forms. We
could change these to use options-messages with the message keys, but it
looks like this was done so the list could be alphabetised based on the
labels and this wouldn't be in alphabetical order if we sorted on the
message keys.

Bug: T278014
Change-Id: I5f59ccf4c167756255952cfbf31a8d7891463e92
(cherry picked from commit 6fa142a4fafc5ff942011ede63030095901fbe45)