Escape rcfilters-filter-* messages on ChangesList pages
The rcfilters-filter-*-label messages are output as raw HTML on
ChangesList pages like RecentChanges and Watchlist. If there is
a wrapping label message, the label is properly escaped, but
when there is not, it is appended as raw HTML. This escapes the
label at output by switching to .text().

Bug: T278058
Change-Id: I7106aedced51343439fc54d5bb91620d8a0362f9
(cherry picked from commit a1dfb13a1972ff9c2afbaf0d3b082d628a6cd9ee)