MergeHistory: Avoid potential DB query for trivial message variation
Title::exists( Title::READ_LATEST ) can lead to querrying the master
database just to decide whether to show a variation of message to the
user, and even after consulting the master database, the check can be
negated by the TitleExists hook (T265775), making the db check useless.

This can be simplified by notifying the caller upfront whether the
source page is deleted, since this info is authoritatively known at
the point of the deletion.

Bug: T278295
Change-Id: Iccf61a06306126c03be302dddfdfeb576217a048